fixed course preview build errors
