+++ app/lib/restapi/endpoints.ts
@@ -26,7 +26,7 @@ export const quizReadUrl = "https://khumla-dev-quiz-read.azurewebsites.net/api";
 
 export const quizWriteUrl ="https://khumla-dev-quiz-write.azurewebsites.net/api";
 
-export const documentWrite = "https://0190-154-0-14-142.ngrok-free.app/api"
+export const documentWrite = "https://e4e7-154-0-14-142.ngrok-free.app/api"
 
 
 export const marksWrite = "https://khumla-dev-marks-write.azurewebsites.net/api"
