+++ app/protected/student/course/course-detail/course-preview.css
